working on readme

+++ src/main.cpp
@@ -1,12 +1,12 @@
 #include <mbed.h>
 #include <Thread.h>
 
-const uint8_t led_mask[] = {1,2,4};
+const int led_mask[] = {1,2,4};
 BusOut leds(LED1,LED2,LED3);
 
 int main()
 {
-    int n = 0;
+    unsigned n = 0;
     while(1)
     {
         leds = led_mask[n%3];
